third_party/WebKit/LayoutTests/http/tests/csslayout/parse-secure-context.html

<!DOCTYPE html>
<meta name="assert" content="This test checks that a layout() function correctly parses on a secure context." />
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="../resources/get-host-info.js"></script>
<style>
#test {
  display: flex;
  display: layout(success);
}
</style>
<div id="test"></div>
<script>
  if (window.location.origin != get_host_info().AUTHENTICATED_ORIGIN) {
    window.location = get_host_info().AUTHENTICATED_ORIGIN + window.location.pathname;
  } else {
    test(function() {
      const element = document.getElementById('test');
      assert_true(window.isSecureContext);
      assert_equals(getComputedStyle(element).display, 'layout(success)');
    });
  }
</script>
